+++ changelog.md
@@ -9,6 +9,8 @@ Codion Change Log
 - UserPreferences.flush() no longer flushes file based preferences.
 - UserPreferences.file() moved to FilePreferences.filePreferences().
 - FilePreferences now caches preferences using a Path instead of filename.
+### is.codion.framework.model
+- EntityEditModel.EditEvents.Inserted, Updated and Deleted added, entity type validated.
 ### is.codion.swing.framework.ui
 - EntitySearchField bug fixed, now uses MessageFormat.format() when formatting result limit message.
 - EntityPopupMenu removed, EntityViewer added.